# [Storage] Add db Seeker to optimize speed & memory when seeking key inside a range (BadgerDB and Pebble)