Currently, `Iterator` is used when we need to get a key within a specified range, where the key is less than or equal to a specified key.  This is very inefficient with both BadgerDB and Pebble.

As one example, `store.VersionBeacons.Highest()` is affected by this inefficiency.

This PR adds a `Seeker` interface which can be used to seek a key that is less than or equal to a specified key within a specified range.  `Seeker` uses db provided seek functions to improve speed and memory use.

This PR also refactors `FindHighestAtOrBelowByPrefix()` to use the new Seeker instead of Iterator.

This optimizes `store.VersionBeacons.Highest()`, which uses `FindHighestAtOrBelowByPrefix()`.

I found this while getting familiar with storage code related to BadgerDB and Pebble (not looking for optimizations).

## Benchmarks using BadgerDB

`FindHighestAtOrBelowByPrefix`

```
                                │ iterator.log  │             seeker.log              │
                                │    sec/op     │   sec/op     vs base                │
FindHighestAtOrBelowByPrefix-12   122.370µ ± 2%   2.458µ ± 3%  -97.99% (p=0.000 n=20)

                                │ iterator.log  │              seeker.log              │
                                │     B/op      │     B/op      vs base                │
FindHighestAtOrBelowByPrefix-12   31.290Ki ± 0%   1.723Ki ± 0%  -94.49% (p=0.000 n=20)

                                │ iterator.log │             seeker.log             │
                                │  allocs/op   │ allocs/op   vs base                │
FindHighestAtOrBelowByPrefix-12    885.00 ± 0%   43.00 ± 0%  -95.14% (p=0.000 n=20)
```

Benchmarks tests used data taken from existing tests.

## Benchmarks using Pebble

`FindHighestAtOrBelowByPrefix`

```
                                │ iterator.log │             seeker.log              │
                                │    sec/op    │   sec/op     vs base                │
FindHighestAtOrBelowByPrefix-12    9.598µ ± 0%   1.591µ ± 2%  -83.42% (p=0.000 n=20)

                                │ iterator.log │            seeker.log             │
                                │     B/op     │    B/op     vs base               │
FindHighestAtOrBelowByPrefix-12     641.0 ± 0%   649.0 ± 0%  +1.25% (p=0.000 n=20)

                                │ iterator.log │           seeker.log           │
                                │  allocs/op   │ allocs/op   vs base            │
FindHighestAtOrBelowByPrefix-12     18.00 ± 0%   18.00 ± 0%  ~ (p=1.000 n=20) ¹
```

Benchmarks tests used data taken from existing tests.