@fxamacker 2025-04-21
It seems this PR [reduces block execution time](https://flow-foundation.slack.com/archives/C07NFGGAGHM/p1744906859358079?thread_ts=1744901261.002229&cid=C07NFGGAGHM) by ~15% :tada:. Thanks @j1010001 for setting up the test environment and comparing block execution times.

NOTE: `store.VersionBeacons.Highest()` is faster with this PR because it uses `FindHighestAtOrBelowByPrefix()`.

@j1010001 This PR also reduces memory allocs when using BadgerDB, so maybe it can help reduce GC pressure on EN (reduce CPU usage of EN overall if GC pressure is reduced enough).
- before: 885 allocs/op
- after: 43 allocs/op

Thanks again for comparing block execution time!